Tests/Unit: move mock class to doubles directory
